main.py

+++ main.py
@@ -15,7 +15,7 @@ def clear():
         os.system('clear')
 
 
-def ip_pinger_download(run_option, return_to_menu=True):
+def ip_pinger_download(run_option, return_to_menu=True, Download_option=True):
     ip_pinger_link = "https://raw.githubusercontent.com/lockieLocks/Tools/main/ip_pinger.py"
     folder_name = "ip_pinger"
     os.makedirs(folder_name, exist_ok=True)
@@ -23,9 +23,12 @@ def ip_pinger_download(run_option, return_to_menu=True):
     try:
         response = requests.get(ip_pinger_link)
         if response.status_code == 200:
-            with open(filename, "wb") as file:
-                file.write(response.content)
-            print("Successfully downloaded ip_pinger.py")
+            if Download_option:
+                with open(filename, "wb") as file:
+                    file.write(response.content)
+                    print("Successfully downloaded ip_pinger.py")
+            else:
+                pass
             if run_option.lower().strip() == 'y':       
                 try:
                     subprocess.run([sys.executable, filename])
@@ -48,7 +51,7 @@ def ip_pinger_download(run_option, return_to_menu=True):
     else:
         return
 
-def b64_download(run_option, return_to_menu=True):
+def b64_download(run_option, return_to_menu=True, Download_option=True):
     b64_link = "https://raw.githubusercontent.com/lockieLocks/Tools/main/b64.py"
     folder_name = "base64"
     os.makedirs(folder_name, exist_ok=True)
@@ -57,9 +60,12 @@ def b64_download(run_option, return_to_menu=True):
     try:
         response = requests.get(b64_link)
         if response.status_code == 200:
-            with open(filename, "wb") as file:
+            if Download_option:
+                with open(filename, "wb") as file:
                     file.write(response.content)
-            print("Successfully downloaded b64.py")
+                    print("Successfully downloaded b64.py")
+            else:
+                pass
             if run_option.lower().strip() == 'y':
                 try:
                     subprocess.run([sys.executable, filename]) 
@@ -82,7 +88,7 @@ def b64_download(run_option, return_to_menu=True):
     else:
         return
 
-def username_lookup_download(run_option, return_to_menu=True):
+def username_lookup_download(run_option, return_to_menu=True, Download_option=True):
     username_lookup_link = "https://raw.githubusercontent.com/lockieLocks/Tools/main/username_checker.py"
     folder_name = "username_lookup"
     os.makedirs(folder_name, exist_ok=True)
@@ -91,9 +97,12 @@ def username_lookup_download(run_option, return_to_menu=True):
     try:
         response = requests.get(username_lookup_link)
         if response.status_code == 200:
-            with open(filename, "wb") as file:
+            if Download_option:
+                with open(filename, "wb") as file:
                     file.write(response.content)
-            print("Successfully downloaded username_lookup.py")
+                    print("Successfully downloaded username_lookup.py")
+            else:
+                pass
             if run_option.lower().strip() == 'y':
                 try:
                     subprocess.run([sys.executable, filename]) 
@@ -116,7 +125,7 @@ def username_lookup_download(run_option, return_to_menu=True):
     else:
         return
 
-def ip_lookup_download(run_option, return_to_menu=True):
+def ip_lookup_download(run_option, return_to_menu=True, Download_option=True):
     ip_lookup_link = "https://raw.githubusercontent.com/lockieLocks/Tools/main/ip_lookup.py"
     folder_name = "ip_lookup"
     os.makedirs(folder_name, exist_ok=True)
@@ -124,9 +133,12 @@ def ip_lookup_download(run_option, return_to_menu=True):
     try:
         response = requests.get(ip_lookup_link)
         if response.status_code == 200:
-            with open(filename, "wb") as file:
-                file.write(response.content)
-                print("Successfully downloaded ip_lookup.py")
+            if Download_option:
+                with open(filename, "wb") as file:
+                    file.write(response.content)
+                    print("Successfully downloaded ip_lookup.py")
+            else:
+                pass
             if run_option.lower().strip() == 'y':
                 try:
                     subprocess.run([sys.executable, filename])
@@ -144,7 +156,7 @@ def ip_lookup_download(run_option, return_to_menu=True):
     except Exception as e:
         print(f"Error >> {e}")
 
-def webhook_mutltitool_download(run_option, return_to_menu=True):
+def webhook_mutltitool_download(run_option, return_to_menu=True, Download_option=True):
     webhook_multitool_link = "https://raw.githubusercontent.com/lockieLocks/Tools/main/webhook_multitool.py"
     folder_name = "webhook_multitool"
     os.makedirs(folder_name, exist_ok=True)
@@ -152,9 +164,12 @@ def webhook_mutltitool_download(run_option, return_to_menu=True):
     try:
         response = requests.get(webhook_multitool_link)
         if response.status_code == 200:
-            with open(filename, "wb") as file:
-                file.write(response.content)
-                print("Successfully downloaded webhook_multitool.py")
+            if Download_option:
+                with open(filename, "wb") as file:
+                    file.write(response.content)
+                    print("Successfully downloaded webhook_multitool.py")
+            else:
+                pass
             if run_option.lower().strip() == 'y':
                 try:
                     subprocess.run([sys.executable, filename])
@@ -166,13 +181,13 @@ def webhook_mutltitool_download(run_option, return_to_menu=True):
                     time.sleep(0.5)
                     tools_menu()
                 else:
-                    return
+                    pass
         else:
-            print("Failed to download Ip_lookup")
+            print("Failed to download webhook_multitool.py")
     except Exception as e:
         print(f"Error >> {e}")
 
-def site_checker(run_option, return_to_menu=True):
+def site_checker(run_option, return_to_menu=True, Download_option=True):
     site_multitool_link = "https://raw.githubusercontent.com/lockieLocks/Tools/main/site_checker.py"
     folder_name = "site_multitool"
     os.makedirs(folder_name, exist_ok=True)
@@ -180,9 +195,12 @@ def site_checker(run_option, return_to_menu=True):
     try:
         response = requests.get(site_multitool_link)
         if response.status_code == 200:
-            with open(filename, "wb") as file:
-                file.write(response.content)
-                print("Successfully downloaded site_checker.py")
+            if Download_option:
+                with open(filename, "wb") as file:
+                    file.write(response.content)
+                    print("Successfully downloaded site_checker.py")
+            else:
+                pass
             if run_option.lower().strip() == 'y':
                 try:
                     subprocess.run([sys.executable, filename])
@@ -200,7 +218,7 @@ def site_checker(run_option, return_to_menu=True):
     except Exception as e:
         print(f"Error >> {e}")
 
-def pwd_strength_checker(run_option, return_to_menu=True):
+def pwd_strength_checker(run_option, return_to_menu=True, Download_option=True):
     pwd_strength_link = "https://raw.githubusercontent.com/lockieLocks/Tools/main/pwd_checker.py"
     folder_name = "password_strength_checker"
     os.makedirs(folder_name, exist_ok=True)
@@ -208,9 +226,12 @@ def pwd_strength_checker(run_option, return_to_menu=True):
     try:
         response = requests.get(pwd_strength_link)
         if response.status_code == 200:
-            with open(filename, "wb") as file:
-                file.write(response.content)
-                print("Successfully downloaded site_checker.py")
+            if Download_option:
+                with open(filename, "wb") as file:
+                    file.write(response.content)
+                    print("Successfully downloaded pwd_strength_checker.py")
+            else:
+                pass
             if run_option.lower().strip() == 'y':
                 try:
                     subprocess.run([sys.executable, filename])
@@ -224,7 +245,41 @@ def pwd_strength_checker(run_option, return_to_menu=True):
                 else:
                     return
         else:
-            print("Failed to download pwd_strength_checker.py")
+            print("Failed to download token_validator.py")
+    except Exception as e:
+        print(f"Error >> {e}")
+
+def token_validator(run_option, return_to_menu=True, Download_option=True):
+    token_validator_link = "https://raw.githubusercontent.com/lockieLocks/Tools/main/token_validator.py"
+    folder_name = "token_validator"
+    os.makedirs(folder_name, exist_ok=True)
+    filename = os.path.join(folder_name, "token_validator.py")
+    file_token = os.path.join(folder_name, "tokens.txt")
+    try:
+        response = requests.get(token_validator_link)
+        if response.status_code == 200:
+            if Download_option:
+                with open(filename, "wb") as file:
+                    file.write(response.content)
+                    print("Successfully downloaded token_validator.py")
+            else:
+                return
+            with open(file_token, "w") as t:
+                t.write("put tokens here twin")
+            if run_option.lower().strip() == 'y':
+                try:
+                    subprocess.run([sys.executable, filename])
+                except Exception as e:
+                    print(f"Error >> {e}")
+            else:
+                if return_to_menu:
+                    print("Returning...")
+                    time.sleep(0.5)
+                    tools_menu()
+                else:
+                    return
+        else:
+            print("Failed to download token_validator.py")
     except Exception as e:
         print(f"Error >> {e}")
 
@@ -233,6 +288,8 @@ def update():
     for name in os.listdir(main_dir):
         path = os.path.join(main_dir, name)
         if os.path.isdir(path):
+            if name.lower() in [".vs", "__pycache__"]:  
+                continue
             shutil.rmtree(path)
             print(f"Deleted All Folders >> {path}")
     print("Redownloading files...")
@@ -244,6 +301,7 @@ def update():
     webhook_mutltitool_download(run_option='n', return_to_menu=False)
     site_checker(run_option='n', return_to_menu=False)
     pwd_strength_checker(run_option='n', return_to_menu=False)
+    token_validator(run_option='n', return_to_menu=False)
     input("\nPress Enter to return...")
     tools_menu()
 
@@ -277,30 +335,30 @@ def tools_menu():
     tools_ascii()
     print("                [------Network Tools------]          [------Utility Tools------]             [------Discord Tools------]")
     print("                     [1] - Ip Lookup                    [4] - B64 Tools                       [7] - Webhook Multitool")
-    print("                     [2] - IP Pinger                    [5] - Username Lookup")
+    print("                     [2] - IP Pinger                    [5] - Username Lookup                 [8] - Token Validator")               
     print("                     [3] - Site Multitool               [6] - Password Strength Checker")
     option = input("Option >> ")
     if option == '1':
         run_option = input("You wanna run the tool? [Y or N] >> ")
-        ip_lookup_download(run_option)
+        ip_lookup_download(run_option, Download_option=False)
     elif option == '2':
         run_option = input("You wanna run the tool? [Y or N] >> ")
-        ip_pinger_download(run_option)
+        ip_pinger_download(run_option, Download_option=False)
     elif option == '3':
         run_option = input("You wanna run the tool? [Y or N] >> ")
-        site_checker(run_option)
+        site_checker(run_option, Download_option=False)
     elif option == '4':
         run_option = input("You wanna run the tool? [Y or N] >> ")
-        b64_download(run_option)
+        b64_download(run_option, Download_option=False)
     elif option == '5':
         run_option = input("You wanna run the tool? [Y or N] >> ")
-        username_lookup_download(run_option)
+        username_lookup_download(run_option, Download_option=False)
     elif option == '6':
         run_option = input("You wanna run the tool? [Y or N] >> ")
-        pwd_strength_checker(run_option)
+        pwd_strength_checker(run_option, Download_option=False)
     elif option == '7':
         run_option = input("You wanna run the tool? [Y or N] >> ")
-        webhook_mutltitool_download(run_option)
+        webhook_mutltitool_download(run_option, Download_option=False)
     elif option == '99':
         update()
     elif option == '404':
@@ -342,27 +400,36 @@ def main():
     clear()
     main_ascii()
     print("\n" + " " * 25 + "3:< Welcome to MAIN MENU of lockies multitool >:3" + " " * 25)
-    print("\n                                   [1]  - Install Discord Tools = Webhook spammer")
+    print("\n                                   [1]  - Install Discord Tools = Webhook spammer, Token Validator")
     print("                                   [2]  - Install Network Tools = IP Lookup, IP Pinger, Site Multitool")
     print("                                   [3]  - Install Utility Tools = Username lookup, B64 multitool, Pwd Strength Checker")
     print("                                   [4]  - Install All tools = All of above")
     print("                                   [99] - Exit")
     option = input("Option >> ")
     if option == '1':
-        run_option = input("You wanna run the tool? [Y or N] >> ")
-        webhook_mutltitool_download(run_option, return_to_menu=False)
+        token_validator(run_option='n', return_to_menu=False)
+        webhook_mutltitool_download(run_option='n', return_to_menu=False)
+        print("Taking you to Utillity menu...")
+        time.sleep(0.5)
+        discord_tools_main()
     elif option == '2':
         ip_lookup_download(run_option='n', return_to_menu=False)
         ip_pinger_download(run_option='n', return_to_menu=False)
         site_checker(run_option='n', return_to_menu=False)
+        print("Taking you to Networking menu...")
+        time.sleep(0.5)
         network_main()
     elif option == '3':
         username_lookup_download(run_option='n', return_to_menu=False)
         b64_download(run_option='n', return_to_menu=False)
         pwd_strength_checker(run_option='n', return_to_menu=False)
+        print("Taking you to Utillity menu...")
+        time.sleep(0.5)
         utility_Tools_main()
     elif option == '4':
         install_all()
+        print("Taking you to Tools menu...")
+        time.sleep(0.5)
         tools_menu()
     elif option.lower().strip() == 't':
         tools_menu()
@@ -401,13 +468,13 @@ def utility_Tools_main():
     option = input("Option >> ")
     if option == '1':
         run_option = input("You wanna run the tool? [Y or N] >> ")
-        username_lookup_download(run_option)
+        username_lookup_download(run_option, Download_option=False)
     elif option == '2':
         run_option = input("You wanna run the tool? [Y or N] >> ")
-        b64_download(run_option)
+        b64_download(run_option, Download_option=False)
     elif option == '3':
         run_option = input("You wanna run the tool? [Y or N] >> ")
-        pwd_strength_checker(run_option)
+        pwd_strength_checker(run_option, Download_option=False)
     elif option == '100':
         print("Returning to main menu...")
         main()
@@ -445,19 +512,55 @@ def network_main():
     option = input("Option >> ")
     if option == '1':
         run_option = input("You wanna run the tool? [Y or N] >> ")
-        ip_lookup_download(run_option)
+        ip_lookup_download(run_option, Download_option=False)
     elif option == '2':
         run_option = input("You wanna run the tool? [Y or N] >> ")
-        ip_pinger_download(run_option)
+        ip_pinger_download(run_option, Download_option=False)
     elif option == '3':
         run_option = input("You wanna run the tool? [Y or N] >> ")
-        site_checker(run_option)
+        site_checker(run_option, Download_option=False)
     elif option == '100':
         main()
     else:
         print("Invalid...")
         time.sleep(0.5)
         network_main()
+
+def discord_ascii():
+    discord_ascii = f""" {Fore.LIGHTGREEN_EX}
+    
+ *******   **                                      **       **********                    **        
+/**////** //                                      /**      /////**///                    /**        
+/**    /** **  ******  *****   ******  ******     /**          /**      ******   ******  /**  ******
+/**    /**/** **////  **///** **////**//**//*  ******          /**     **////** **////** /** **//// 
+/**    /**/**//***** /**  // /**   /** /** /  **///**          /**    /**   /**/**   /** /**//***** 
+/**    ** /** /////**/**   **/**   /** /**   /**  /**          /**    /**   /**/**   /** /** /////**
+/*******  /** ****** //***** //****** /***   //******          /**    //****** //******  *** ****** 
+///////   // //////   /////   //////  ///     //////           //      //////   //////  /// //////  
+                                        b - Return to main menu
+"""
+    print(discord_ascii)
+
+def discord_tools_main():
+    clear()
+    discord_ascii()
+    print("                        [1] - Run Webhook Multitool")
+    print("                        [2] - Run Token Validator")
+    option = input("Option >> ")
+    if option == '1':
+        run_option = input("You wanna run the tool? [Y or N] >> ")
+        webhook_mutltitool_download(run_option, Download_option=False)
+    elif option == '2':
+        run_option = input("You wanna run the tool? [Y or N] >> ")
+        token_validator(run_option, Download_option=False)
+    elif option.strip() == 'b':
+        print("Taking you to main menu...")
+        time.sleep(0.5)
+        main()
+    else:
+        print("Invalid...")
+        time.sleep(0.5)
+        discord_tools_main()
 
 if __name__ == '__main__':
     main()
